containers-verus + egraph: rustfmt, clear every warning, and pin the triggers
Three kinds of hygiene the campaign left behind, in one commit because the
files carry all three and hunk-level attribution would be invented rather
than recovered.

Formatting: 35 files. `main` is fmt-clean at zero files, so every one of these
came from this campaign and `cargo fmt --all -- --check` has been failing.

Warnings: 21, of which only some were clippy lints. CI sets
`RUSTFLAGS: -D warnings`, so the dead code and unused assignments failed it
too. Most are mechanical. Three are worth naming. `CacheFrame::saved_len` was
written at three sites and never read, and its own comment already said
restore needs nothing from it, so removing the field also drops a
`self.nodes.len().as_usize()` per mark. `mark_and_compact` and
`mark_and_compact_sorted` keep an allow rather than being deleted, so their
contracts stay proved. In `RunCol::restore_to` the trailing `g` increment is
read only by the proof (`pairs.subrange(0, g as int)`), which rustc cannot
see.

Two allows are crate-level because verus rejects what the lint suggests: the
`vec!` macro expands to a let expression and collapsing nested `if let`s
produces a let-chain, and verus supports neither ("does not yet support the
following Rust feature: let expressions"). Measured, not assumed: rewriting
the three `Vec::new`-then-`push` sites as `vec![]` and letting rustfix
collapse `sync_group::poke` both broke verification, and both are reverted
here. They sit at crate level because the `verus!` macro drops
statement-level allows.

Triggers: 18 quantifiers were taking automatically chosen triggers, which
verus prints when it has low confidence in the choice. Fifteen now carry
`#![auto]`, which accepts the trigger verus already picked and therefore
changes no proof behaviour. Three in `hinted_arena` carried a `#[trigger]`
inside `&(...)`, where verus does not accept it and silently fell back to
auto; those move to binder position as `#![trigger self.view()[j]]`.

Verification is unchanged at 2164 verified, 0 errors, before and after. Tests
are 1861 passed across 147 binaries, 0 failures. fmt and clippy are both at
zero.

+++ containers-conformance/tests/choose_mode_criterion.rs
@@ -15,17 +15,29 @@ use verus::diff_compress::choose_mode;
 // returns IndexRunsSorted (its cost is the sorted run count, which the sorted
 // encoder achieves), but the write-order IndexRuns is the same family.
 fn is_runs(m: CompressionMode) -> bool {
-    matches!(m, CompressionMode::IndexRuns | CompressionMode::IndexRunsSorted)
+    matches!(
+        m,
+        CompressionMode::IndexRuns | CompressionMode::IndexRunsSorted
+    )
 }
-fn is_dict(m: CompressionMode) -> bool { matches!(m, CompressionMode::ValueDict) }
-fn is_none(m: CompressionMode) -> bool { matches!(m, CompressionMode::None) }
+fn is_dict(m: CompressionMode) -> bool {
+    matches!(m, CompressionMode::ValueDict)
+}
+fn is_none(m: CompressionMode) -> bool {
+    matches!(m, CompressionMode::None)
+}
 
 #[test]
 fn clustered_indices_pick_index_major() {
     // One contiguous run of 500 cells, each a distinct value. Index-major drops
     // the index column (1 run); value-major cannot help (all values distinct).
-    let diffs: Vec<(u32, u32)> = (0..500u32).map(|i| (i.wrapping_mul(2654435761), i)).collect();
-    assert!(is_runs(choose_mode::<u32, u32>(&diffs)), "clustered indices -> IndexRuns");
+    let diffs: Vec<(u32, u32)> = (0..500u32)
+        .map(|i| (i.wrapping_mul(2654435761), i))
+        .collect();
+    assert!(
+        is_runs(choose_mode::<u32, u32>(&diffs)),
+        "clustered indices -> IndexRuns"
+    );
 }
 
 #[test]
@@ -36,8 +48,9 @@ fn repeated_values_scattered_indices_pick_value_major() {
     // = 2141 < plain 4000, so the honest selector picks ValueDict. (At the old usize
     // codes this was 6016 > plain and it wrongly picked None — the packed-codes fix
     // is exactly what lets the selector choose value-major here.)
-    let diffs: Vec<(u32, u32)> =
-        (0..500u32).map(|i| ((i % 4) as u32, i.wrapping_mul(7) % 100_000)).collect();
+    let diffs: Vec<(u32, u32)> = (0..500u32)
+        .map(|i| ((i % 4), i.wrapping_mul(7) % 100_000))
+        .collect();
     assert!(
         is_dict(choose_mode::<u32, u32>(&diffs)),
         "few distinct, scattered, narrow codes -> ValueDict (now a win)"
@@ -51,7 +64,10 @@ fn scattered_unique_picks_plain() {
     let diffs: Vec<(u32, u32)> = (0..500u32)
         .map(|i| (i.wrapping_mul(2654435761), i.wrapping_mul(7) % 100_000))
         .collect();
-    assert!(is_none(choose_mode::<u32, u32>(&diffs)), "scattered + unique -> None (plain)");
+    assert!(
+        is_none(choose_mode::<u32, u32>(&diffs)),
+        "scattered + unique -> None (plain)"
+    );
 }
 
 #[test]
